+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 This project uses [SemVer](https://semver.org/) for versioning. Its public APIs, runtime support, and documented file locations won't change incompatibly outside of major versions (once version 1.0.0 has been released). There may be breaking changes in minor releases before 1.0.0 and will be noted in these release notes.
 
+## 0.3.0
+
+_released `TBD`_
+
+- add a `universal-test-runner` command which has info about the package. The main test runner still lives in `t`.
+- remove custom file caching logic
+
 ## 0.2.0
 
 _released `2023-06-18`_